refactor: move all exported testing into single file
